#3386 - Improved DIAM serialization
- Rename package to make checkstyle happy
